feat: production hardening — security, observability, code quality, tests, CI
Security:
- Float → Numeric(12,2) for all financial fields (users, orders, transactions)
- Payment webhook idempotency via PaymentLog table + UNIQUE constraint
- 2FA god_mode fail-closed (Redis unavailable → 503 instead of bypass)
- Custom Redis rate limiter replacing broken slowapi (fixes nginx proxy crash)
- Input validation: Query params with ge/le, file extension fix, batch max_length=50
- WebSocket authentication via Telegram initData validation
- Auth window reduced from 24h to 1h

Observability:
- Replaced 10 debug print() with structured logger calls
- Added logging to 15+ silent except Exception: pass clauses
- Sentry SDK integration (backend config + init)
- Health endpoint with PostgreSQL + Redis connectivity checks
- OpenAPI docs enabled in debug mode

Code quality:
- ESLint flat config + Prettier for frontend
- Ruff linter + pyproject.toml for Python
- Production console.log wrapped in import.meta.env.DEV guards
- SectionErrorBoundary default fallback UI instead of null
- Archive/unarchive idempotency guards
- Removed unused variable in CreateOrderPage

Database:
- Alembic migration: indexes (status, created_at, work_type), CHECK constraints
  (progress 0-100, price >= 0, discount 0-100), payment_logs table

Frontend:
- Inline styles → Tailwind: OrderDetailPageV8 (257), OrderChatPage (26), HomePage (6)

Tests & CI:
- Backend tests: test_auth.py, test_rate_limit.py, test_orders_api.py
- CI pipeline: test job (ruff, pytest, tsc, npm test) gates deploy

+++ database/models/transactions.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, String, func
+from sqlalchemy import BigInteger, DateTime, Numeric, ForeignKey, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database.db import Base
@@ -15,7 +16,7 @@ class BalanceTransaction(Base):
     user_id: Mapped[int] = mapped_column(
         BigInteger, ForeignKey("users.telegram_id", ondelete="CASCADE"), index=True
     )
-    amount: Mapped[float] = mapped_column(Float, nullable=False)  # TODO: migrate Float→Numeric(12,2)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     type: Mapped[str] = mapped_column(String(20), nullable=False)  # credit / debit
     reason: Mapped[str] = mapped_column(String(50), nullable=False)
     description: Mapped[str | None] = mapped_column(String(255), nullable=True)
